+++ examples/all.sh
@@ -7,4 +7,5 @@ echo Running all ARK Ledger Examples..
 echo ""
 sh ${TEST_DIR}/legacy.sh
 sh ${TEST_DIR}/v2.sh
+sh ${TEST_DIR}/message.sh
 echo ""