+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+-   added message signing features ([#40])
 -   implement v2 Transactions ([#27])
 -   added VendorField display support ([#29])
 -   added build options and documentation ([#32])
@@ -22,4 +23,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [#27]: https://github.com/ArkEcosystem/ledger/pull/27
 [#29]: https://github.com/ArkEcosystem/ledger/pull/29
 [#32]: https://github.com/ArkEcosystem/ledger/pull/32
+[#40]: https://github.com/ArkEcosystem/ledger/pull/40
 [unreleased]: https://github.com/ArkEcosystem/ledger/compare/master...develop